Check if assets folder exists before copy

+++ scripts/src/plh-copy-app-data.ts
@@ -7,11 +7,11 @@ const DATA_INPUT_FOLDER = path.join(__dirname, "./plh-data-convert/output");
 const ASSETS_INPUT_FOLDER = path.join(__dirname, "./gdrive-download/output/plh_assets");
 const APP_DATA_DIR = `../src/data`;
 const APP_PLH_ASSETS_DIR = "../src/assets/plh_assets";
-fs.ensureDirSync(APP_DATA_DIR);
-fs.emptyDirSync(APP_DATA_DIR);
 
 /** A simple script to copy data exported from gdrive and processed for plh into the app data folder */
 export function main(doAssetFolderCheck = true) {
+  fs.ensureDirSync(APP_DATA_DIR);
+  fs.emptyDirSync(APP_DATA_DIR);
   if (doAssetFolderCheck) {
     fs.ensureDirSync(APP_PLH_ASSETS_DIR);
     fs.emptyDirSync(APP_PLH_ASSETS_DIR);
@@ -24,7 +24,9 @@ export function main(doAssetFolderCheck = true) {
     const appOutputTs = generateAppTsOutput(localTsFile);
     fs.writeFileSync(`${APP_DATA_DIR}/${flow_type}.ts`, appOutputTs);
   }
-  fs.copySync(ASSETS_INPUT_FOLDER, APP_PLH_ASSETS_DIR);
+  if (fs.existsSync(ASSETS_INPUT_FOLDER)) {
+    fs.copySync(ASSETS_INPUT_FOLDER, APP_PLH_ASSETS_DIR);
+  }
   console.log(chalk.green("Data Copied to App"));
 }
